ENH Nimbler cal_kl implementation
These matrices can be quite large and the previous implementation
created a lot of intermediate ones

Also, log(v1/v2) == log(v1)/log(v2), so execute the log before the
broadcast
