+++ .github/workflows/semibin_test.yml
@@ -48,6 +48,7 @@ jobs:
         python -m pytest ./test/test_train.py
         python -m pytest ./test/test_bin.py
         python -m pytest ./test/test_utils.py
+        python -m pytest ./test/test_cal_kl.py
     - name: Test with SemiBin command
       shell: bash -l {0}
       run: |
